pyproject.toml
[project]
name = "agentic-platform"
version = "0.1.0"
description = "Progressive LangGraph learning platform — Research Agent capstone"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # LangChain / LangGraph core
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-sqlite>=1.0.0",

    # LLM providers
    "langchain-anthropic>=0.2.0",

    # Tool / search integrations
    "tavily-python>=0.3.0",
    "duckduckgo-search>=6.0.0",
    "arxiv>=2.1.0",
    "wikipedia>=1.4.0",

    # HTTP & async
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",

    # Data & validation
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",

    # Observability
    "langsmith>=0.1.0",
    "structlog>=24.1.0",

    # CLI
    "typer>=0.12.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "ipython>=8.24.0",
    "jupyter>=1.0.0",
]

[project.scripts]
research = "agentic_platform.research_agent.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_platform"]

[tool.ruff]
src = ["src"]
line-length = 100
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
